refactor(mobile): extract Files screen into embeddable panel
MobileFileExplorerPanel (props-driven, embedded mode drops full-screen chrome
and swaps back->close); route file is now a thin wrapper. Behavior unchanged
full-screen (U3).
